+++ src/app/music-and-tabs/page.js
@@ -1,15 +1,15 @@
-import { projectCategories } from '@/const/projectCategories';
+import { PROJECTS_SUB_NAVIGATION } from '@/const/SUB_NAVIGATION';
 import PageTitle from '@/components/PageTitle/PageTitle';
-import CategoriesNavigation from '@/components/CategoriesNavigation/CategoriesNavigation';
+import SubNavigation from '@/components/SubNavigation/SubNavigation';
 import PostsGrid from '@/components/PostsGrid/PostsGrid';
 
 const MusicAndTabs = () => {
   return (
     <main className="container">
       <PageTitle title="Music & Tabs." />
-      <CategoriesNavigation categories={projectCategories} showAll={true} />
+      <SubNavigation navigationItems={PROJECTS_SUB_NAVIGATION} showAll={true} />
       <PostsGrid
-        type="project"
+        postType="project"
         category="Music & Tabs"
         postsToShow={6}
         postsToLoad={6}