+++ src/app/reading-this-year/page.js
@@ -1,15 +1,20 @@
-import { libraryLinks } from '@/const/libraryLinks';
-import { readingThisYear } from '@/const/readingThisYear';
+import { LIBRARY_SUB_NAVIGATION } from '@/const/SUB_NAVIGATION';
 import PageTitle from '@/components/PageTitle/PageTitle';
-import CategoriesNavigation from '@/components/CategoriesNavigation/CategoriesNavigation';
-import BooksGrid from '@/components/BooksGrid/BooksGrid';
+import SubNavigation from '@/components/SubNavigation/SubNavigation';
+import PostsGrid from '@/components/PostsGrid/PostsGrid';
+PostsGrid;
 
 const ReadingThisYear = () => {
   return (
     <main className="container">
       <PageTitle title="Library: Reading this Year" />
-      <CategoriesNavigation categories={libraryLinks} showAll={false} />
-      <BooksGrid content={readingThisYear} postsToShow={12} postsToLoad={6} />
+      <SubNavigation navigationItems={LIBRARY_SUB_NAVIGATION} showAll={false} />
+      <PostsGrid
+        postType="book"
+        category="Reading this Year"
+        postsToShow={12}
+        postsToLoad={6}
+      />
     </main>
   );
 };