+++ src/const/SUB_NAVIGATION.js
@@ -1,0 +1,25 @@
+export const PROJECTS_SUB_NAVIGATION = [
+  {
+    title: 'Illustrations',
+    url: '/illustrations',
+  },
+  {
+    title: 'Music & Tabs',
+    url: '/music-and-tabs',
+  },
+];
+
+export const LIBRARY_SUB_NAVIGATION = [
+  {
+    title: 'Reading this Year',
+    url: '/reading-this-year',
+  },
+  {
+    title: 'Want to Read',
+    url: '/want-to-read',
+  },
+  {
+    title: 'Years of Reading',
+    url: '/years-of-reading',
+  },
+];